Support giftwrap package installation
This change is similar to the recent change to support installing
from source via a common dependent role. The difference here is that
we are installing from a giftwrap artifact. This artifact includes
a virtualenv of the project being installed & the sample etc configs
for that build, but does not deal with installing services into
$PATH.

As we want multiple versions to be able to live side-by-side, these
packages merely provide the bits. It is up to ursula to stich them
into a running set of services.  To do that, we install the package
and ensure that '/opt/openstack/current/' exists. This will be the
point from which all other activities (ie. dealing with rootwrap,
update-alternatives, and dealing with upgrades) will take place.