Update Horizon to play nice with packaging
With packaging comes a different way to deliver Horizon. Previously,
we had insertted configs directly into the horizon virtualenv that
we built in-place. Since we are no longer building in-place, and
dropping files into the paths owned by a package is pretty nasty,
update horizon to handle this case.

Now the config is placed in /etc/openstack_dashboard/local_settings.py.